backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI/ML
openai>=1.6.1
anthropic==0.7.7
langchain==0.0.335
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# OCR
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Caching and storage
redis==5.0.1
boto3==1.34.0  # For AWS S3

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
celery==5.3.4  # For background tasks
python-magic==0.4.27

# Development
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1 